requirements.txt
Django==1.9.7
django-geoposition==0.2.3
django-tastypie==0.13.3
Pillow==3.3.0
